+++ XVim2/XcodeHeader/IDEKit/IDEWorkspaceWindow.h
@@ -49,7 +49,7 @@
 // Remaining properties
 @property(readonly, copy) NSString *debugDescription;
 @property(readonly, copy) NSString *description;
-@property(readonly) unsigned long long hash;
+@property(readonly) NSUInteger hash;
 @property(readonly) Class superclass;
 
 @end
